For 2.0: Bump minimum Ruby version to 2.7.0
Ruby 2.6 is long since out of maintenance so I think we should bump the minimum version for 2.0.

There's still a big Bullet Train app on 2.7, so we can't just push this to 3.0 yet.

We can just have this PR open until just before 2.0 and then merge, so it makes any needed 1.x releases easier.